+++ CHANGES.md
@@ -7,6 +7,7 @@
 
 ### Changed
 
+- Improve `Vcs.Tree.sexp_of_t` to help with debugging.
 - Rename `git_cli` library to `vcs_git_cli` for consistency.
 - Remove type parameter for `Vcs.Tree.Node_kind` (simplify interface).
 - Renamed constructors for root nodes in vcs trees (`Init` => `Root`).